intake-agent/requirements.txt
# AI/ML Libraries
anthropic>=0.25.0

# Database
psycopg2-binary>=2.9.9

# Cloud Storage
boto3>=1.34.0
google-cloud-storage>=2.10.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0

# Web Framework
flask>=3.0.0

# File Monitoring
watchdog>=4.0.0

# Document Processing
pdfplumber>=0.10.0
extract-msg>=0.48.0

# Configuration & Utilities
python-dotenv>=1.0.0
requests>=2.31.0